+++ model/test/fetch_test.dart
@@ -1,4 +1,3 @@
-import 'package:get_it/get_it.dart';
 import 'package:oghref_model/buffer_parser.dart';
 import 'package:oghref_model/src/parser/open_graph.dart';
 import 'package:oghref_model/src/parser/twitter_card.dart';
@@ -9,22 +8,22 @@ final Uri resourseUri = Uri.parse(
 
 void main() {
   setUpAll(() {
-    GetIt.I.registerSingleton<MetaFetch>(MetaFetch.forTest()
+    MetaFetch.instance = MetaFetch.forTest()
       ..register(const OpenGraphPropertyParser())
       ..register(const TwitterCardPropertyParser())
-      ..primaryPrefix = "og");
+      ..primaryPrefix = "og";
   });
 
   test("Test parse under HTTPS", () async {
     final parsed =
-        await GetIt.I<MetaFetch>().fetchFromHttp(resourseUri.resolve("1.html"));
+        await MetaFetch.instance.fetchFromHttp(resourseUri.resolve("1.html"));
 
     expect(parsed.title, equals("Sample 1"));
     expect(parsed.images.first.width, equals(400.0));
   });
   test("Ignore subproperties content", () async {
     final parsed =
-        await GetIt.I<MetaFetch>().fetchFromHttp(resourseUri.resolve("2.html"));
+        await MetaFetch.instance.fetchFromHttp(resourseUri.resolve("2.html"));
 
     expect(parsed.images.first.width, isNull);
     expect(parsed.images.first.height, isNull);
@@ -33,7 +32,7 @@ void main() {
 
   test("Unaccept JavaScript generated HTML element", () async {
     final parsed =
-        await GetIt.I<MetaFetch>().fetchFromHttp(resourseUri.resolve("3.html"));
+        await MetaFetch.instance.fetchFromHttp(resourseUri.resolve("3.html"));
 
     expect(parsed.title, isNull);
     expect(parsed.url, isNull);